requirements.txt
tweepy==4.14.0
requests==2.31.0
beautifulsoup4==4.10.0
python-dotenv==0.19.0
ratelimit==2.2.1
datetime==5.4